pyproject.toml
[tool.poetry]
name = "discord-reargparse"
version = "0.0.1"
description = "RegEx based argument parsing for commands in discord.py"
license = "MIT"
authors = ["nkpro2000sr <srnaveen2k@yahoo.com>"]
maintainers = ["nkpro2000sr <srnaveen2k@yahoo.com>"]
readme = "README.md"
repository = "https://github.com/nkpro2000sr/discord-reargparse"
packages = [
    { include = "discord_reargparse"},
]
keywords = ["discord","regex","argparse","reargparse"]

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-asyncio = "^0.12.0"
dpytest = "^0.0.20"

[tool.black]
line-length = 99

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
